+++ test/help/format-command.test.ts
@@ -339,7 +339,7 @@ FLAGS
                      <options: a|b|c>`)
     })
 
-    it('should output flag enum with helpValue', async () => {
+    it('should output flag enum with string helpValue', async () => {
       const cmd = await makeLoadable(
         makeCommandClass({
           id: 'apps:create',
@@ -358,6 +358,69 @@ FLAGS
 
 FLAGS
   --myenum=a|b|c`)
+    })
+
+    it('should output flag enum with array helpValue', async () => {
+      const cmd = await makeLoadable(
+        makeCommandClass({
+          id: 'apps:create',
+          flags: {
+            myenum: flags.string({
+              options: myEnumValues,
+              helpValue: myEnumValues,
+            }),
+          },
+        }),
+      )
+
+      const output = help.formatCommand(cmd)
+      expect(output).to.equal(`USAGE
+  $ oclif apps:create [--myenum a b c]
+
+FLAGS
+  --myenum=a b c`)
+    })
+
+    it('should output string helpValue in usage string', async () => {
+      const cmd = await makeLoadable(
+        makeCommandClass({
+          id: 'apps:create',
+          flags: {
+            files: flags.string({
+              helpValue: '<input-json>|<input-xml>',
+              multiple: true,
+            }),
+          },
+        }),
+      )
+
+      const output = help.formatCommand(cmd)
+      expect(output).to.equal(`USAGE
+  $ oclif apps:create [--files <input-json>|<input-xml>...]
+
+FLAGS
+  --files=<input-json>|<input-xml>...`)
+    })
+
+    it('should output array helpValue in usage string', async () => {
+      const cmd = await makeLoadable(
+        makeCommandClass({
+          id: 'apps:create',
+          flags: {
+            files: flags.string({
+              helpValue: ['<input-json>', '<input-xml>'],
+              multiple: true,
+            }),
+          },
+        }),
+      )
+
+      const output = help.formatCommand(cmd)
+      expect(output).to.equal(`USAGE
+  $ oclif apps:create [--files <input-json>... <input-xml>...]
+
+FLAGS
+  --files=<input-json>... <input-xml>...`)
     })
 
     it('should output with default flag options', async () => {
@@ -576,6 +639,53 @@ ARGUMENTS
       const output = help.formatCommand(cmd)
       expect(output).to.equal(`USAGE
   $ oclif apps:create`)
+    })
+
+    it('should output usage with hardcoded command', async () => {
+      const cmd = await makeLoadable(
+        makeCommandClass({
+          id: 'apps:create',
+          usage: ['apps:create'],
+        }),
+      )
+      const output = help.formatCommand(cmd)
+      expect(output).to.equal(`USAGE
+  $ oclif apps:create`)
+    })
+
+    it('should output default usage for single letter command', async () => {
+      const cmd = await makeLoadable(
+        makeCommandClass({
+          id: 'a',
+          flags: {
+            'a-flag': flags.string({char: 'a', options: ['a', 'aa', 'aaa']}),
+          },
+        }),
+      )
+      const output = help.formatCommand(cmd)
+      expect(output).to.equal(`USAGE
+  $ oclif a [-a a|aa|aaa]
+
+FLAGS
+  -a, --a-flag=<option>  <options: a|aa|aaa>`)
+    })
+
+    it('should output usage for single letter command', async () => {
+      const cmd = await makeLoadable(
+        makeCommandClass({
+          id: 'a',
+          flags: {
+            'a-flag': flags.string({char: 'a', options: ['a', 'aa', 'aaa']}),
+          },
+          usage: 'a [-a a|aa|aaa]',
+        }),
+      )
+      const output = help.formatCommand(cmd)
+      expect(output).to.equal(`USAGE
+  $ oclif a [-a a|aa|aaa]
+
+FLAGS
+  -a, --a-flag=<option>  <options: a|aa|aaa>`)
     })
   })